requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.34
alembic==1.13.2
PyJWT==2.9.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
psycopg[binary]==3.2.3
